replace lambda aggregate script with batch version
