--- data-serving/scripts/aggregate-batch/aggregate/__init__.py
@@ -1,1 +1,0 @@
-boto3 = "1.16.60"
