--- data-serving/scripts/aggregate-batch/aggregate/variants.json
@@ -1,1 +1,0 @@
-{"South Africa": {"_id": "South Africa", "casecount_p1": 0.0, "casecount_b1351": 887.0, "casecount": 0, "lat": -29.00034095, "long": 25.08390093}, "United Kingdom": {"_id": "United Kingdom", "casecount_p1": 0.0, "casecount_b1351": 128.0, "casecount": 0, "lat": 54.12387156, "long": -2.86563164}, "France": {"_id": "France", "casecount_p1": 4.0, "casecount_b1351": 61.0, "casecount": 0, "lat": 42.17344011, "long": -2.76172945}, "Belgium": {"_id": "Belgium", "casecount_p1": 0.0, "casecount_b1351": 48.0, "casecount": 0, "lat": 50.63981576, "long": 4.64065114}, "Netherlands": {"_id": "Netherlands", "casecount_p1": 2.0, "casecount_b1351": 32.0, "casecount": 0, "lat": 52.1007899, "long": 5.28144793}, "Switzerland": {"_id": "Switzerland", "casecount_p1": 2.0, "casecount_b1351": 23.0, "casecount": 0, "lat": 46.79785878, "long": 8.20867471}, "Botswana": {"_id": "Botswana", "casecount_p1": 0.0, "casecount_b1351": 22.0, "casecount": 0, "lat": -22.18403213, "long": 23.79853368}, "Mozambique": {"_id": "Mozambique", "casecount_p1": 0.0, "casecount_b1351": 20.0, "casecount": 0, "lat": -17.27381643, "long": 35.53367543}, "Australia": {"_id": "Australia", "casecount_p1": 0.0, "casecount_b1351": 16.0, "casecount": 0, "lat": -25.73288704, "long": 134.49100008}, "Germany": {"_id": "Germany", "casecount_p1": 1.0, "casecount_b1351": 15.0, "casecount": 0, "lat": 51.10698181, "long": 10.38578051}, "United States": {"_id": "United States", "casecount_p1": 3.0, "casecount_b1351": 13.0, "casecount": 0, "lat": 45.6795472, "long": -112.4616737}, "Israel": {"_id": "Israel", "casecount_p1": 0.0, "casecount_b1351": 14.0, "casecount": 0, "lat": 31.46110101, "long": 35.00444693}, "Ireland": {"_id": "Ireland", "casecount_p1": 0.0, "casecount_b1351": 9.0, "casecount": 0, "lat": 53.1754487, "long": -8.13793569}, "New Zealand": {"_id": "New Zealand", "casecount_p1": 0.0, "casecount_b1351": 7.0, "casecount": 0, "lat": -41.81113557, "long": 171.48492347}, "Canada": {"_id": "Canada", "casecount_p1": 1.0, "casecount_b1351": 7.0, "casecount": 0, "lat": 61.36206324, "long": -98.30777028}, "Japan": {"_id": "Japan", "casecount_p1": 4.0, "casecount_b1351": 7.0, "casecount": 0, "lat": 37.59230135, "long": 138.03089558}, "United Arab Emirates": {"_id": "United Arab Emirates", "casecount_p1": 0.0, "casecount_b1351": 6.0, "casecount": 0, "lat": 23.90528188, "long": 54.3001671}, "Turkey": {"_id": "Turkey", "casecount_p1": 1.0, "casecount_b1351": 5.0, "casecount": 0, "lat": 39.0616029, "long": 35.16895346}, "Denmark": {"_id": "Denmark", "casecount_p1": 0.0, "casecount_b1351": 5.0, "casecount": 0, "lat": 55.98125296, "long": 10.02800992}, "Austria": {"_id": "Austria", "casecount_p1": 0.0, "casecount_b1351": 4.0, "casecount": 0, "lat": 47.58549439, "long": 14.1264761}, "Spain": {"_id": "Spain", "casecount_p1": 1.0, "casecount_b1351": 4.0, "casecount": 0, "lat": 40.24448698, "long": -3.64755047}, "Zimbabwe": {"_id": "Zimbabwe", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": -19.00420419, "long": 29.8514412}, "Finland": {"_id": "Finland", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 64.49884603, "long": 26.2746656}, "Portugal": {"_id": "Portugal", "casecount_p1": 1.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 39.59550671, "long": -8.50104361}, "South Korea": {"_id": "South Korea", "casecount_p1": 1.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 36.38523983, "long": 127.83916086}, "Bangladesh": {"_id": "Bangladesh", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 23.86731158, "long": 90.23812743}, "Panama": {"_id": "Panama", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 8.51750797, "long": -80.11915156}, "Luxembourg": {"_id": "Luxembourg", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 49.76725361, "long": 6.07182201}, "Sweden": {"_id": "Sweden", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 62.77966519, "long": 16.74558049}, "Norway": {"_id": "Norway", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 68.75015572, "long": 15.34834656}, "Zambia": {"_id": "Zambia", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": -13.45824152, "long": 27.77475946}, "Kenya": {"_id": "Kenya", "casecount_p1": 0.0, "casecount_b1351": 2.0, "casecount": 0, "lat": 0.59988022, "long": 37.79593973}, "Taiwan": {"_id": "Taiwan", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 23.7539928, "long": 120.95427281}, "Ghana": {"_id": "Ghana", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 7.95345644, "long": -1.21676566}, "Vietnam": {"_id": "Vietnam", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 16.6460167, "long": 106.29914698}, "Gambia": {"_id": "Gambia", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 13.44965244, "long": -15.39601295}, "China": {"_id": "China", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 36.56176546, "long": 103.81907349}, "Greece": {"_id": "Greece", "casecount_p1": 0.0, "casecount_b1351": 1.0, "casecount": 0, "lat": 39.07469623, "long": 22.95555794}, "Brazil": {"_id": "Brazil", "casecount_p1": 82.0, "casecount_b1351": 0.0, "casecount": 0, "lat": -10.78777702, "long": -53.09783113}, "Italy": {"_id": "Italy", "casecount_p1": 4.0, "casecount_b1351": 0.0, "casecount": 0, "lat": 42.79662641, "long": 12.07001339}, "Peru": {"_id": "Peru", "casecount_p1": 3.0, "casecount_b1351": 0.0, "casecount": 0, "lat": -9.15280381, "long": -74.38242685}, "Colombia": {"_id": "Colombia", "casecount_p1": 2.0, "casecount_b1351": 0.0, "casecount": 0, "lat": 3.91383431, "long": -73.08114582}, "Venezuela": {"_id": "Venezuela", "casecount_p1": 1.0, "casecount_b1351": 0.0, "casecount": 0, "lat": 7.12422421, "long": -66.18184123}, "Faroe Islands": {"_id": "Faroe Islands", "casecount_p1": 1.0, "casecount_b1351": 0.0, "casecount": 0, "lat": 61.50507, "long": -6.76918}, "Argentina": {"_id": "Argentina", "casecount_p1": 1.0, "casecount_b1351": 0.0, "casecount": 0, "lat": -35.3813488, "long": -65.17980692}}
