--- data-serving/scripts/aggregate/aggregate/requirements.txt
@@ -1,3 +1,0 @@
-boto3==1.16.60
-pymongo==3.11.2
-dnspython==2.1.0
