--- data-serving/scripts/aggregate/template.yaml
@@ -1,56 +1,0 @@
-AWSTemplateFormatVersion: "2010-09-09"
-Transform: AWS::Serverless-2016-10-31
-Description: AWS Lambda functions for Global.health Data Export
-
-Globals:
-  Function:
-    Runtime: python3.8
-    Timeout: 900
-    EventInvokeConfig:
-      MaximumRetryAttempts: 0
-    Environment:
-      Variables:
-        MONGO_USERNAME: exportservice
-        MONGO_PASSWORD: EXPORT_DB_PASSWORD
-
-Parameters:
-  SecurityGroupIds:
-    Type: CommaDelimitedList
-    Default: sg-0f3446d2b82eff09a
-  SubnetIDs:
-    Type: CommaDelimitedList
-    Description: The list of SubnetIDs in your Virtual Private Cloud (VPC)
-    Default: subnet-01cdb8802584b0891,subnet-0ce56af866d39d69e,subnet-02ac7023a699cfce3,subnet-060e2152a9beb6300,subnet-00253e04dfd3b0269,subnet-0efa6c09f2e0ce2e1
-  EFSpath:
-    Type: String
-    Default: /mnt/efs
-  Version:
-    Type: Number
-    Default: 1
-
-Resources:
-  DataExportAggregateFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: aggregate/
-      Handler: app.lambda_handler
-      Description: Generate aggregate data for map
-      Events:
-        AggSchedule:
-          Type: Schedule
-          Properties:
-            Schedule: 'cron(9 8,20 ? * * *)'
-            Description: Triggers function after prune-uploads
-            Enabled: True
-      Layers:
-        - arn:aws:lambda:us-east-1:612888738066:layer:awsdatawrangler:1
-      Policies:
-        - S3CrudPolicy:
-            BucketName: covid-19-aggregates
-        - AWSLambda_FullAccess
-
-  Bucket:
-    Type: 'AWS::S3::Bucket'
-    Properties:
-      BucketName: global-dot-health-data-export-bucket-2
-
